refactor(cli): move the sort group into its own module
Second step of splitting `cli/main.py`, and the pilot for the ten group
modules that follow. `sort` was picked because it is contiguous, has no
embedded business logic and uses only helpers that already have a home.

The pattern the remaining groups copy:

* `geoparquet_io/cli/commands/<group>.py` declares a standalone
  `@click.group()` (not `@cli.group()`) and its subcommands, and imports
  helpers from `cli/_shared.py` and `cli/decorators.py` only.
* `cli/main.py` imports the group object and attaches it with an explicit
  `cli.add_command(<group>)`, in one registration block under the root group.

Registration is deliberately not done by the group module: `@cli.group()`
there would need `cli.main`, which imports the group -- a cycle. Keeping every
attachment in one block in `main.py` also means `grep add_command` lists the
whole tree, which import-time self-registration would not.

The 330 moved lines are byte-identical to their previous form apart from the
group's decorator line. `main.py` re-exports `sort`, so
`from geoparquet_io.cli.main import sort` still resolves.

Also updated:

* `test_write_memory_forwarding` and `test_cli_api_call_parity_scaffold` patch
  the CLI's `*_impl` alias in the module that owns the command, so the four
  sort cases now name `cli.commands.sort`. Both files carry the module
  explicitly, per case, so the next group move is a one-line edit.
* the `doc-sync`, `validate-claude-md` and `check-api-for-cli` hook file
  patterns, so a change confined to `cli/commands/` still triggers them.

`tests/data/cli_surface.json` is unchanged: `gpio --help`, `gpio sort --help`
and `gpio sort hilbert --help` render identically, and a real `gpio sort
hilbert` run produces a byte-identical output file.
